# perf: inline binary search