--- a/node.go
+++ b/node.go
@@ -69,37 +69,77 @@ func (n *Node[T]) replaceEdge(e edge[T]) {
 }
 
 func (n *Node[T]) getEdge(label byte) (int, *Node[T]) {
-	num := len(n.edges)
-	idx := sort.Search(num, func(i int) bool {
-		return n.edges[i].label >= label
-	})
-	if idx < num && n.edges[idx].label == label {
-		return idx, n.edges[idx].node
+	// Binary search for the edge.
+	// This is a performance-critical section hence the manual binary search.
+	// Using sort.Search is ~50% slower due to the function call overhead and
+	// lack of early return.
+	for low, high := 0, len(n.edges); low < high; {
+		mid := int(uint(low+high) >> 1)
+		// low ≤ mid < high
+		if n.edges[mid].label == label {
+			return mid, n.edges[mid].node
+		}
+
+		if n.edges[mid].label < label {
+			low = mid + 1
+		} else {
+			high = mid
+		}
 	}
+
+	// No match.
 	return -1, nil
 }
 
 func (n *Node[T]) getLowerBoundEdge(label byte) (int, *Node[T]) {
-	num := len(n.edges)
-	idx := sort.Search(num, func(i int) bool {
-		return n.edges[i].label >= label
-	})
-	// we want lower bound behavior so return even if it's not an exact match
-	if idx < num {
-		return idx, n.edges[idx].node
+	// Binary search for the edge.
+	// This is a performance-critical section hence the manual binary search.
+	// Using sort.Search is ~50% slower due to the function call overhead and
+	// lack of early return.
+	low, high := 0, len(n.edges)
+	for low < high {
+		mid := int(uint(low+high) >> 1)
+		// low ≤ mid < high
+		if n.edges[mid].label == label {
+			return mid, n.edges[mid].node
+		}
+
+		if n.edges[mid].label < label {
+			low = mid + 1
+		} else {
+			high = mid
+		}
 	}
+
+	// we want lower bound behavior so return even if it's not an exact match
+	if low < len(n.edges) {
+		return low, n.edges[low].node
+	}
+
 	return -1, nil
 }
 
 func (n *Node[T]) delEdge(label byte) {
-	num := len(n.edges)
-	idx := sort.Search(num, func(i int) bool {
-		return n.edges[i].label >= label
-	})
-	if idx < num && n.edges[idx].label == label {
-		copy(n.edges[idx:], n.edges[idx+1:])
-		n.edges[len(n.edges)-1] = edge[T]{}
-		n.edges = n.edges[:len(n.edges)-1]
+	// Binary search for the edge.
+	// This is a performance-critical section hence the manual binary search.
+	// Using sort.Search is ~50% slower due to the function call overhead and
+	// lack of early return.
+	low, high := 0, len(n.edges)
+	for low < high {
+		mid := int(uint(low+high) >> 1)
+		// low ≤ mid < high
+		if n.edges[mid].label == label {
+			copy(n.edges[mid:], n.edges[mid+1:])
+			n.edges[len(n.edges)-1] = edge[T]{}
+			n.edges = n.edges[:len(n.edges)-1]
+			return
+		}
+
+		if n.edges[mid].label < label {
+			low = mid + 1
+		} else {
+			high = mid
+		}
 	}
 }
 
@@ -162,7 +202,7 @@ func (n *Node[T]) LongestPrefix(k []byte) ([]byte, T, bool) {
 			break
 		}
 
-		// Consume the search prefix
+		// Consume the search prefix.
 		if bytes.HasPrefix(search, n.prefix) {
 			search = search[len(n.prefix):]
 		} else {