… go.sum
… testdata/endpoints.txt
… testdata/uuids.txt
… testdata/words.txt